+++ include/algorithms/PlanePca.h
@@ -15,20 +15,20 @@
 using namespace Eigen;
 using namespace std;
 
-typedef Matrix<float, 3, 3> Mat33;
+typedef Matrix<double, 3, 3> Mat33;
 
 class PlanePca {
 private:
-    static void pointsVectorToMatrix(const vector<Vector3f> &pointsVector, MatrixXf &matrix);
+    static void pointsVectorToMatrix(const vector<Vector3d> &pointsVector, MatrixXf &matrix);
 
-    static Mat33 computeCovMatrix(const vector<Vector3f> &pointsVector, const Vector3f &mean);
+    static Mat33 computeCovMatrix(const vector<Vector3d> &pointsVector, const Vector3d &mean);
 
-    static Vector3f computeMean(const vector<Vector3f> &pointsVector);
+    static Vector3d computeMean(const vector<Vector3d> &pointsVector);
 
-    static Plane computePlane(const vector<Vector3f> &pointsVector, const Mat& colorImage, const ImageCoords& imageCoords);
+    static Plane computePlane(const vector<Vector3d> &pointsVector, const Mat& colorImage, const ImageCoords& imageCoords);
 
 public:
-    static Plane getPlane(const vector<Vector3f> &pointsVector, const Mat& colorImage, const ImageCoords& imageCoords);
+    static Plane getPlane(const vector<Vector3d> &pointsVector, const Mat& colorImage, const ImageCoords& imageCoords);
 };
