+++ include/algorithms/PlaneRansac.h
@@ -17,12 +17,12 @@ using namespace std;
 class PlaneRansac {
 private:
     static const int MAX_ITERATIONS_NUM = 10000;
-    static constexpr float INLAIERS_PERCENT_CONSENSUS = 0.8f;
-    static constexpr float MAX_INLARIES_POINT_PLANE_DISTANCE = 1.0f;
-    static void getRandom3Points(const vector<Vector3f> &pointsVector, Vector3f *random3Points);
-    static Plane computePlane(const vector<Vector3f> &pointsVector, const Mat &colorImage );
+    static constexpr double INLAIERS_PERCENT_CONSENSUS = 0.8f;
+    static constexpr double MAX_INLARIES_POINT_PLANE_DISTANCE = 1.0f;
+    static void getRandom3Points(const vector<Vector3d> &pointsVector, Vector3d *random3Points);
+    static Plane computePlane(const vector<Vector3d> &pointsVector, const Mat &colorImage );
 public:
-    static Plane getPlane(const vector<Vector3f> &pointsVector, const Mat &colorImage );
+    static Plane getPlane(const vector<Vector3d> &pointsVector, const Mat &colorImage );
 };
