+++ include/models/PointCloud.h
@@ -15,10 +15,10 @@
 class PointCloud {
 private:
     std::vector<Point3D> points3D;
-    std::vector<Eigen::Vector3f> points;
+    std::vector<Eigen::Vector3d> points;
     Eigen::Matrix<double, 3, 3> PHCPModel;
 
-    void getPoint(unsigned int u, unsigned int v, float depth, Eigen::Vector3d &point3D);
+    void getPoint(unsigned int u, unsigned int v, double depth, Eigen::Vector3d &point3D);
 
 public:
     PointCloud();
@@ -27,7 +27,7 @@ public:
     void push_back(Point3D point3D);
     void clear();
     const std::vector<Point3D> &getPoints3D() const;
-    const std::vector<Eigen::Vector3f> &getPoints() const;
+    const std::vector<Eigen::Vector3d> &getPoints() const;
 };
