+++ src/models/Plane.cpp
@@ -10,7 +10,7 @@
 
 Plane::Plane() {}
 
-Plane::Plane(Vector3f point1, Vector3f point2, Vector3f point3, const Mat& colorImage) {
+Plane::Plane(Vector3d point1, Vector3d point2, Vector3d point3, const Mat& colorImage) {
     color = HSVColor(colorImage);
 }
 
@@ -18,7 +18,7 @@ Plane::Plane(std::array<Vector3d, 3>, const Mat& colorImage) {
     color = HSVColor(colorImage);
 }
 
-Plane::Plane(Vector3f normalVec, Vector3f point, const Mat& colorImage, const vector<Vector3f> &points, const ImageCoords &imageCoords) {
+Plane::Plane(Vector3d normalVec, Vector3d point, const Mat& colorImage, const vector<Vector3d> &points, const ImageCoords &imageCoords) {
     planeNormalVec = normalVec;
     planeNormalVec.normalize();
     A = normalVec(0);
@@ -38,7 +38,7 @@ Plane::Plane(Vector3f normalVec, Vector3f point, const Mat& colorImage, const ve
     valid = true;
 }
 
-Plane::Plane(Vector3f normalVec, float D, vector<Vector3f> points, vector<ImageCoords> imageCoordsVec, HSVColor color){
+Plane::Plane(Vector3d normalVec, double D, vector<Vector3d> points, vector<ImageCoords> imageCoordsVec, HSVColor color){
     this->planeNormalVec = normalVec;
     this->D = D;
     this->A = normalVec(0);
@@ -57,7 +57,7 @@ Plane::Plane(Vector3f normalVec, float D, vector<Vector3f> points, vector<ImageC
     valid = true;
 }
 
-Plane::Plane(Vector3f normalVec, float D){
+Plane::Plane(Vector3d normalVec, double D){
     planeNormalVec = normalVec;
     planeNormalVec.normalize();
     A = normalVec(0);
@@ -74,31 +74,31 @@ Plane::Plane(Vector3f normalVec, float D){
     valid = true;
 }
 
-float Plane::getA() const {
+double Plane::getA() const {
     return A;
 }
 
-float Plane::getB() const {
+double Plane::getB() const {
     return B;
 }
 
-float Plane::getC() const {
+double Plane::getC() const {
     return C;
 }
 
-float Plane::getD() const {
+double Plane::getD() const {
     return D;
 }
 
-float Plane::getDistanceFromPoint(Vector3f point) {
+double Plane::getDistanceFromPoint(Vector3d point) {
     return std::abs(A * point(0) + B * point(1) + C * point(2) + D)
            / sqrtf(powf(A, 2) + powf(B, 2) + powf(C, 2));
 }
 
-void Plane::computePlaneEquation(Vector3f point1, Vector3f point2, Vector3f point3) {
-    Eigen::Vector3f v = point1 - point2;
-    Eigen::Vector3f w = point1 - point3;
-    Eigen::Vector3f planeParameters = v.cross(w);
+void Plane::computePlaneEquation(Vector3d point1, Vector3d point2, Vector3d point3) {
+    Eigen::Vector3d v = point1 - point2;
+    Eigen::Vector3d w = point1 - point3;
+    Eigen::Vector3d planeParameters = v.cross(w);
     planeParameters.normalize();
     A = planeParameters(0);
     B = planeParameters(1);
@@ -125,31 +125,31 @@ const HSVColor &Plane::getColor() const {
     return color;
 }
 
-Vector3f Plane::computePointOnPlaneFromTwoCoordinate(float firstCoordinate, float secondCoordinate) const {
-    Vector3f pointToReturn;
+Vector3d Plane::computePointOnPlaneFromTwoCoordinate(double firstCoordinate, double secondCoordinate) const {
+    Vector3d pointToReturn;
     if(A == 0 && B == 0 && C == 0){
         throw std::runtime_error("Plane equation error!");
     } else if( A == 0 && B == 0){
-        float z = D / B;
+        double z = D / B;
         pointToReturn(0) = firstCoordinate;
         pointToReturn(1) = secondCoordinate;
         pointToReturn(2) = z;
     } else if(A == 0 && C == 0){
-        float y = D / B;
+        double y = D / B;
         pointToReturn(0) = firstCoordinate;
         pointToReturn(1) = y;
         pointToReturn(2) = secondCoordinate;
     } else if(B==0 && C==0){
-        float x = D / A;
+        double x = D / A;
         pointToReturn(0) = x;
         pointToReturn(1) = firstCoordinate;
     } else if(C == 0){
-        float y = (D - A * firstCoordinate) / B;
+        double y = (D - A * firstCoordinate) / B;
         pointToReturn(0) = firstCoordinate;
         pointToReturn(1) = y;
         pointToReturn(2) = secondCoordinate;
     } else{
-        float z = (D - A * firstCoordinate - B * secondCoordinate) / C;
+        double z = (D - A * firstCoordinate - B * secondCoordinate) / C;
         pointToReturn(0) = firstCoordinate;
         pointToReturn(1) = secondCoordinate;
         pointToReturn(2) = z;
@@ -157,7 +157,7 @@ Vector3f Plane::computePointOnPlaneFromTwoCoordinate(float firstCoordinate, floa
     return  pointToReturn;
 }
 
-Vector3f Plane::getPlaneNormalVec() const{
+Vector3d Plane::getPlaneNormalVec() const{
     return planeNormalVec;
 }
 
@@ -169,11 +169,11 @@ void Plane::setImageCoords(const ImageCoords &imageCoords) {
     Plane::imageCoordsVec.push_back(imageCoords);
 }
 
-const vector<Vector3f> &Plane::getPoints() const {
+const vector<Vector3d> &Plane::getPoints() const {
     return points;
 }
 
-void Plane::setPoints(const vector<Vector3f> &points) {
+void Plane::setPoints(const vector<Vector3d> &points) {
     Plane::points = points;
 }
 
@@ -181,13 +181,13 @@ unsigned int Plane::getNumberOfPoints() const {
     return (unsigned int)points.size();
 }
 
-float Plane::getAngleBetweenTwoPlanes(const Plane &plane) const {
-    Eigen::Vector3f planeNormalVec = plane.getPlaneNormalVec();
+double Plane::getAngleBetweenTwoPlanes(const Plane &plane) const {
+    Eigen::Vector3d planeNormalVec = plane.getPlaneNormalVec();
 
-    float angleCos = this->planeNormalVec.dot(planeNormalVec) / this->planeNormalVec.norm() / planeNormalVec.norm();
+    double angleCos = this->planeNormalVec.dot(planeNormalVec) / this->planeNormalVec.norm() / planeNormalVec.norm();
     if(angleCos < -1) angleCos = -1.0f;
     if(angleCos > 1) angleCos = 1.0f;
-    float angle = acosf(angleCos)*180.0f/(float)M_PI;
+    double angle = acosf(angleCos)*180.0f/(double)M_PI;
 /*    if(angle > 90.0f){
         angle = 180.0f - angle;
     }*/
@@ -198,7 +198,7 @@ void Plane::setColor(const HSVColor &color) {
     Plane::color = color;
 }
 
-void Plane::insertPoints(vector<Vector3f> points){
+void Plane::insertPoints(vector<Vector3d> points){
     this->points.insert(this->points.end(), points.begin(), points.end());
 }
 
@@ -217,7 +217,7 @@ const vector<ImageCoords> &Plane::getImageCoordsVec() const {
     return imageCoordsVec;
 }
 
-Vector3f Plane::getCentralPoint() const {
+Vector3d Plane::getCentralPoint() const {
     return points[(points.size() - 1) / 2];
 }
